+++ lib/src/screens/dashboard/dashboard_page.dart
@@ -3,9 +3,6 @@ import 'package:cake_wallet/di.dart';
 import 'package:cake_wallet/entities/main_actions.dart';
 import 'package:cake_wallet/src/screens/dashboard/desktop_widgets/desktop_wallet_selection_dropdown.dart';
 import 'package:cake_wallet/src/screens/dashboard/desktop_widgets/desktop_dashboard_view.dart';
-import 'package:cake_wallet/src/screens/dashboard/desktop_widgets/desktop_sidebar/side_menu.dart';
-import 'package:cake_wallet/src/screens/dashboard/desktop_widgets/desktop_sidebar/side_menu_controller.dart';
-import 'package:cake_wallet/src/screens/dashboard/desktop_widgets/desktop_sidebar/side_menu_item.dart';
 import 'package:cake_wallet/src/screens/dashboard/widgets/market_place_page.dart';
 import 'package:cake_wallet/generated/i18n.dart';
 import 'package:cake_wallet/routes.dart';
@@ -27,121 +24,9 @@ import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:mobx/mobx.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
 import 'package:cake_wallet/main.dart';
-import 'package:cake_wallet/router.dart' as Router;
 
-class DashboardPage extends StatefulWidget {
+class DashboardPage extends BasePage {
   DashboardPage({
-    required this.balancePage,
-    required this.walletViewModel,
-    required this.addressListViewModel,
-  });
-
-  final BalancePage balancePage;
-  final DashboardViewModel walletViewModel;
-  final WalletAddressListViewModel addressListViewModel;
-
-  @override
-  State<DashboardPage> createState() => _DashboardPageState();
-}
-
-class _DashboardPageState extends State<DashboardPage> {
-  final page = PageController();
-  final sideMenu = SideMenuController();
-  @override
-  void initState() {
-    SideMenuGlobal.controller = sideMenu;
-    sideMenu.addListener((p0) {
-      page.jumpToPage(p0);
-    });
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (ResponsiveLayoutUtil.instance.isMobile(context)) {
-      return Scaffold(
-        body: _DashboardPage(
-          balancePage: widget.balancePage,
-          walletViewModel: widget.walletViewModel,
-          addressListViewModel: widget.addressListViewModel,
-        ),
-      );
-    }
-
-    return Scaffold(
-      body: Container(
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.start,
-          children: [
-            SideMenu(
-              topItems: [
-                SideMenuItem(
-                  iconPath: 'assets/images/wallet_outline.png',
-                  priority: 0,
-                  onTap: (page, _) {
-                    sideMenu.changePage(page);
-                  },
-                ),
-              ],
-              bottomItems: [
-                SideMenuItem(
-                  iconPath: 'assets/images/support_icon.png',
-                  priority: 1,
-                  onTap: (page, _) {
-                    sideMenu.changePage(page);
-                  },
-                ),
-                SideMenuItem(
-                  iconPath: 'assets/images/settings_outline.png',
-                  priority: 2,
-                  onTap: (page, _) {
-                    sideMenu.changePage(page);
-                  },
-                ),
-              ],
-            ),
-            Expanded(
-              child: PageView(
-                controller: page,
-                physics: NeverScrollableScrollPhysics(),
-                children: [
-                  _DashboardPage(
-                    balancePage: widget.balancePage,
-                    walletViewModel: widget.walletViewModel,
-                    addressListViewModel: widget.addressListViewModel,
-                  ),
-                  Container(
-                    child: Navigator(
-                      initialRoute: Routes.support,
-                      onGenerateRoute: (settings) => Router.createRoute(settings),
-                      onGenerateInitialRoutes: (NavigatorState navigator, String initialRouteName) {
-                        return [
-                          navigator.widget.onGenerateRoute!(RouteSettings(name: initialRouteName))!
-                        ];
-                      },
-                    ),
-                  ),
-                  Navigator(
-                    initialRoute: Routes.desktop_settings_page,
-                    onGenerateRoute: (settings) => Router.createRoute(settings),
-                    onGenerateInitialRoutes: (NavigatorState navigator, String initialRouteName) {
-                      return [
-                        navigator.widget.onGenerateRoute!(RouteSettings(name: initialRouteName))!
-                      ];
-                    },
-                  ),
-                ],
-              ),
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _DashboardPage extends BasePage {
-  _DashboardPage({
     required this.balancePage,
     required this.walletViewModel,
     required this.addressListViewModel,
